Lista para comenzar a implementar tabla en Buscar
