+++ package.json
@@ -13,7 +13,8 @@
     "firebase": "^4.9.0",
     "moment": "^2.20.1",
     "vue": "^2.5.2",
-    "vue-router": "^3.0.1"
+    "vue-router": "^3.0.1",
+    "vuefire": "^1.4.5"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.2",
